fix(techdocs): fix page re-renders
