+++ plugins/techdocs/src/reader/components/TechDocsReaderPage/context.tsx
@@ -15,92 +15,42 @@
  */
 
 import React, {
-  createContext,
+  ReactNode,
+  memo,
   Dispatch,
-  PropsWithChildren,
   SetStateAction,
+  createContext,
   useContext,
   useState,
 } from 'react';
 import useAsync, { AsyncState } from 'react-use/lib/useAsync';
 
-import { Page } from '@backstage/core-components';
 import { useApi } from '@backstage/core-plugin-api';
 import { CompoundEntityRef } from '@backstage/catalog-model';
 
 import { techdocsApiRef } from '../../../api';
 import { TechDocsEntityMetadata, TechDocsMetadata } from '../../../types';
 
-type PropsWithEntityName<T = {}> = T &
-  PropsWithChildren<{ entityName: CompoundEntityRef }>;
-
-const initialContextValue = {
-  loading: true,
-  error: undefined,
-  value: undefined,
-};
-
-const TechDocsMetadataContext =
-  createContext<AsyncState<TechDocsMetadata>>(initialContextValue);
-
-export const TechDocsMetadataProvider = ({
-  entityName,
-  children,
-}: PropsWithEntityName) => {
-  const techdocsApi = useApi(techdocsApiRef);
-
-  const value = useAsync(async () => {
-    return techdocsApi.getTechDocsMetadata(entityName);
-  }, [entityName]);
-
-  return (
-    <TechDocsMetadataContext.Provider value={value}>
-      {children}
-    </TechDocsMetadataContext.Provider>
-  );
-};
-
-/**
- * Hook for use within TechDocs addons to retrieve TechDocs Metadata for the
- * current TechDocs site.
- * @public
- */
-export const useTechDocsMetadata = () => {
-  return useContext(TechDocsMetadataContext);
-};
-
-const TechDocsEntityContext =
-  createContext<AsyncState<TechDocsEntityMetadata>>(initialContextValue);
-
-export const TechDocsEntityProvider = ({
-  entityName,
-  children,
-}: PropsWithEntityName) => {
-  const techdocsApi = useApi(techdocsApiRef);
-
-  const value = useAsync(async () => {
-    return techdocsApi.getEntityMetadata(entityName);
-  }, [entityName]);
-
-  return (
-    <TechDocsEntityContext.Provider value={value}>
-      {children}
-    </TechDocsEntityContext.Provider>
-  );
-};
-
-/**
- * Hook for use within TechDocs addons to retrieve Entity Metadata for the
- * current TechDocs site.
- * @public
- */
-export const useEntityMetadata = () => {
-  return useContext(TechDocsEntityContext);
+const areEntityNamesEqual = (
+  prevEntityName: CompoundEntityRef,
+  nextEntityName: CompoundEntityRef,
+) => {
+  if (prevEntityName.kind !== nextEntityName.kind) {
+    return false;
+  }
+  if (prevEntityName.name !== nextEntityName.name) {
+    return false;
+  }
+  if (prevEntityName.namespace !== nextEntityName.namespace) {
+    return false;
+  }
+  return true;
 };
 
 export type TechDocsReaderPageValue = {
-  path: string;
+  metadata: AsyncState<TechDocsMetadata>;
   entityName: CompoundEntityRef;
+  entityMetadata: AsyncState<TechDocsEntityMetadata>;
   shadowRoot?: ShadowRoot;
   setShadowRoot: Dispatch<SetStateAction<ShadowRoot | undefined>>;
   title: string;
@@ -110,12 +60,13 @@ export type TechDocsReaderPageValue = {
 };
 
 export const defaultTechDocsReaderPageValue: TechDocsReaderPageValue = {
-  path: '',
   title: '',
-  setTitle: () => {},
   subtitle: '',
+  setTitle: () => {},
   setSubtitle: () => {},
   setShadowRoot: () => {},
+  metadata: { loading: true },
+  entityMetadata: { loading: true },
   entityName: { kind: '', name: '', namespace: '' },
 };
 
@@ -127,48 +78,74 @@ export const useTechDocsReaderPage = () => {
   return useContext(TechDocsReaderPageContext);
 };
 
-type TechDocsReaderPageProviderProps = PropsWithEntityName<{
-  path?: string;
-}>;
+type TechDocsReaderPageProviderRenderFunction = (
+  value: TechDocsReaderPageValue,
+) => JSX.Element;
 
-export const TechDocsReaderPageProvider = ({
-  path = '',
-  entityName,
-  children,
-}: TechDocsReaderPageProviderProps) => {
-  const { value: entityMetadataValue } = useEntityMetadata();
-  const { value: techdocsMetadataValue } = useTechDocsMetadata();
+type TechDocsReaderPageProviderProps = {
+  entityName: CompoundEntityRef;
+  children: TechDocsReaderPageProviderRenderFunction | ReactNode;
+};
 
-  const [title, setTitle] = useState(defaultTechDocsReaderPageValue.title);
-  const [subtitle, setSubtitle] = useState(
-    defaultTechDocsReaderPageValue.subtitle,
-  );
-  const [shadowRoot, setShadowRoot] = useState<ShadowRoot | undefined>(
-    defaultTechDocsReaderPageValue.shadowRoot,
-  );
+export const TechDocsReaderPageProvider = memo(
+  ({ entityName, children }: TechDocsReaderPageProviderProps) => {
+    const techdocsApi = useApi(techdocsApiRef);
 
-  const value = {
-    path,
-    entityName,
-    shadowRoot,
-    setShadowRoot,
-    title,
-    setTitle,
-    subtitle,
-    setSubtitle,
-  };
+    const metadata = useAsync(async () => {
+      return techdocsApi.getTechDocsMetadata(entityName);
+    }, [entityName]);
 
-  return (
-    <TechDocsReaderPageContext.Provider value={value}>
-      <Page themeId="documentation">
-        {children instanceof Function
-          ? children({
-              entityRef: entityName,
-              entityMetadataValue,
-              techdocsMetadataValue,
-            })
-          : children}
-      </Page>
-    </TechDocsReaderPageContext.Provider>
-  );
+    const entityMetadata = useAsync(async () => {
+      return techdocsApi.getEntityMetadata(entityName);
+    }, [entityName]);
+
+    const [title, setTitle] = useState(defaultTechDocsReaderPageValue.title);
+    const [subtitle, setSubtitle] = useState(
+      defaultTechDocsReaderPageValue.subtitle,
+    );
+    const [shadowRoot, setShadowRoot] = useState<ShadowRoot | undefined>(
+      defaultTechDocsReaderPageValue.shadowRoot,
+    );
+
+    const value = {
+      metadata,
+      entityName,
+      entityMetadata,
+      shadowRoot,
+      setShadowRoot,
+      title,
+      setTitle,
+      subtitle,
+      setSubtitle,
+    };
+
+    return (
+      <TechDocsReaderPageContext.Provider value={value}>
+        {children instanceof Function ? children(value) : children}
+      </TechDocsReaderPageContext.Provider>
+    );
+  },
+  (prevProps, nextProps) => {
+    return areEntityNamesEqual(prevProps.entityName, nextProps.entityName);
+  },
+);
+
+/**
+ * Hook for use within TechDocs addons to retrieve Entity Metadata for the
+ * current TechDocs site.
+ * @public
+ */
+export const useEntityMetadata = () => {
+  const { entityMetadata } = useTechDocsReaderPage();
+  return entityMetadata;
+};
+
+/**
+ * Hook for use within TechDocs addons to retrieve TechDocs Metadata for the
+ * current TechDocs site.
+ * @public
+ */
+export const useTechDocsMetadata = () => {
+  const { metadata } = useTechDocsReaderPage();
+  return metadata;
 };
