+++ plugins/techdocs/src/reader/components/TechDocsReaderPageContent/index.ts
@@ -16,3 +16,4 @@
 
 export { TechDocsReaderPageContent } from './TechDocsReaderPageContent';
 export * from './context';
+export * from './dom';
